+++ modules/src/main/java/com/atlassian/plugin/connect/modules/beans/builder/ConnectAddonBeanBuilder.java
@@ -192,6 +192,8 @@ public class ConnectAddonBeanBuilder<T extends ConnectAddonBeanBuilder, B extend
         }
     }
 
+    public String getKey() { return key; }
+
     public AuthenticationBean getAuthentication()
     {
         return authentication;