+++ plugin/src/test/resources/test-main.js
@@ -21,6 +21,7 @@ requirejs.config({
     'iframe/host/_status_helper': '../src/main/resources/js/iframe/host/_status_helper',
     'iframe/host/_dollar': '../src/main/resources/js/iframe/host/_dollar',
     'iframe/host/content': '../src/main/resources/js/iframe/host/content',
+    'iframe/host/history': '../src/main/resources/js/iframe/host/history',
     'dialog/main': '../src/main/resources/js/dialog/main',
     'dialog/button': '../src/main/resources/js/dialog/button',
     'dialog/dialog-factory': '../src/main/resources/js/dialog/dialog-factory',
@@ -28,6 +29,7 @@ requirejs.config({
     'inline-dialog/simple': '../src/main/resources/js/inline-dialog/simple',
     'confluence/macro/editor': '../src/main/resources/js/confluence/macro/editor',
     'jira/event': '../src/main/resources/js/jira/event',
+    'jira/workflow-post-function': '../src/main/resources/js/jira/workflow-post-function',
     'messages/main': '../src/main/resources/js/messages/main',
     // shared
     'iframe/_amd': '../src/main/resources/js/iframe/_amd',
@@ -61,6 +63,12 @@ requirejs.config({
         'aui-atlassian',
         'iframe/_amd',
         'iframe/_ui-params'
+        ]
+    },
+    'iframe/host/history': {
+        deps: [
+        'iframe/host/_dollar',
+        'iframe/_uri'
         ]
     },
     'iframe/_base64': {
@@ -116,6 +124,11 @@ requirejs.config({
         ]
     },
     'jira/event': {
+        deps: [
+        'iframe/host/_dollar'
+        ]
+    },
+    'jira/workflow-post-function': {
         deps: [
         'iframe/host/_dollar'
         ]