+++ docs/public/resources/release-notes.md
@@ -3,6 +3,7 @@
 
 ## [1.1.0](../release-notes/1-1-rc.html)
 * Support for managing cookies through javascript. Check out the [cookie api](../javascript/module-cookie.html)
+* Support for managing browser history through javascript. Check out the [history api](../javascript/module-history.html)
 
 ## [1.0.1](../release-notes/1-0-1.html)
 * Allow POST method for "screens/addToDefault/{fieldId}"